Fix all the things.
1. Fix #16.  Pipe emits 'data', and switching into old-mode while piping
is not allowed.

2. Move all the magic fields to a '_readableOptions' member, so that they're
less likely to be collided with in extension classes.

3. Fix #18.  That was just a normal bug anyway, not any kind of serious
API issue.

4. Fix #17.  Switch into old-mode on pause()/resume()

5. Fix #8.  If you do read(10) and only 9 bytes are available, and the stream
is not ending, then it returns null, not 9 bytes.  It'll never return fewer
bytes than asked for.  If you don't specify a number, then it still gives
you the full buffer.

This is not a very atomic commit.  But then again, this whole repo is a wip.
So, deal with it.