fix compatibility with MongoDB CursorStream
Since v0.0.4 Readable::wrap() requires a length property on
received chunks before emitting a readable event.

A MongoDB CursorStream emits objects in it's data event that lack a
length property resulting in no readable events when wrapped around
the new 0.0.4 wrapper.

Not sure about the best solution or the importance of a correct
length value, but this simple patch reverts back to 0.0.3 behavior.