Exclude `stream` module in browsers
React Native complains about the call to `require('st' + 'ream')` in `readable.js` with `Unable to resolve module 'stream' from 'bla bla bla/readable-stream/readable.js'`. The `stream` module isn't needed for browsers as far as I can tell, and keeping it as-is means that any module relying on streams within the context of react-native is broken unless you jump through a long line of hoops.

So far I've had issues with this in the context of the `mqtt` and `bl` modules, however there's loads of modules out there that will also have these problems.