Fix circular dependency
At the time `require('inherits')(Duplex, Readable)` is called,
Readable is hoisted but not yet defined due to a circular dependency.

Because the only uses of Duplex by the base classes are checking for
if they themselves are a Duplex, we can replace typeof with a boolean.

This boolean actually already exists in Writable, so it makes sense
to use it elsewhere to solve the circular dependency.